// pokeapi/src/app/pokemons/interface/pokemon.interfaces.ts
export interface PokemonRow {
  name: string;
  first_types: string;
  second_types: string;
  attack: number | undefined;
  image: string | undefined;
}